Remove old memory document cache
