+++ src/GraphQL.ApiTests/GraphQL.MemoryCache.approved.txt
@@ -1,14 +1,9 @@
 namespace GraphQL.Caching
 {
-    public class MemoryDocumentCache : GraphQL.Caching.IDocumentCache
+    public class MemoryDocumentCache : GraphQL.Caching.IDocumentCache, System.IDisposable
     {
-        public MemoryDocumentCache(int maxTotalQueryLength) { }
-        public GraphQL.Language.AST.Document this[string query] { get; set; }
-    }
-    public class MsMemoryDocumentCache : GraphQL.Caching.IDocumentCache, System.IDisposable
-    {
-        public MsMemoryDocumentCache(Microsoft.Extensions.Options.IOptions<Microsoft.Extensions.Caching.Memory.MemoryCacheOptions> options, System.TimeSpan slidingExpiration) { }
-        public MsMemoryDocumentCache(int maxTotalQueryLength, System.TimeSpan slidingExpiration) { }
+        public MemoryDocumentCache(Microsoft.Extensions.Options.IOptions<Microsoft.Extensions.Caching.Memory.MemoryCacheOptions> options, System.TimeSpan slidingExpiration) { }
+        public MemoryDocumentCache(int maxTotalQueryLength, System.TimeSpan slidingExpiration) { }
         public GraphQL.Language.AST.Document this[string query] { get; set; }
         public void Dispose() { }
     }
